+++ kovtun.alexander/T2/io_scope_guard.hpp
